Move LocationResolver to ProfileResourceFactory in profile.
