--- droid-command-line/src/main/java/uk/gov/nationalarchives/droid/command/action/LocationResolver.java
@@ -1,67 +1,0 @@
-/**
- * Copyright (c) 2016, The National Archives <pronom@nationalarchives.gov.uk>
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following
- * conditions are met:
- *
- *  * Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- *
- *  * Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in the
- *    documentation and/or other materials provided with the distribution.
- *
- *  * Neither the name of the The National Archives nor the
- *    names of its contributors may be used to endorse or promote products
- *    derived from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
- * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
- * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
- * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
- * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
- * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
- * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package uk.gov.nationalarchives.droid.command.action;
-
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-
-import uk.gov.nationalarchives.droid.profile.AbstractProfileResource;
-import uk.gov.nationalarchives.droid.profile.DirectoryProfileResource;
-import uk.gov.nationalarchives.droid.profile.FileProfileResource;
-
-/**
- * @author rflitcroft
- *
- */
-public class LocationResolver {
-
-    /**
-     * Resolves a location string to a profile resource.
-     * @param location the resources location string
-     * @param recursive if the resource should be recursed
-     * @return a new profile resource
-     */
-    public AbstractProfileResource getResource(String location, boolean recursive) {
-        
-        final Path f = Paths.get(location);
-        if (Files.isRegularFile(f)) {
-            return new FileProfileResource(f);
-        } else if (Files.isDirectory(f)) {
-            return new DirectoryProfileResource(f, recursive);
-        } else {
-            throw new IllegalArgumentException(
-                    String.format("Unknown location [%s]", location));
-        }
-        
-    }
-}
